Extract AddressToFunction and AddressToBlock in separate file
